+++ packages/sdk-react/src/components/payables/PayablesTable/types.ts
@@ -1,11 +1,12 @@
 import { components } from '@/api';
+import { API } from '@/api/client';
 
 import {
   FILTER_TYPE_CREATED_AT,
+  FILTER_TYPE_SUMMARY_CARD,
   FILTER_TYPE_DUE_DATE,
   FILTER_TYPE_SEARCH,
   FILTER_TYPE_STATUS,
-  FILTER_TYPE_OVERDUE,
 } from './consts';
 
 export type Sort = {
@@ -13,13 +14,13 @@ export type Sort = {
   order: 'asc' | 'desc';
 };
 
-export type FilterTypes = {
-  [FILTER_TYPE_SEARCH]?: string | null;
-  [FILTER_TYPE_STATUS]?: components['schemas']['PayableStateEnum'] | null;
-  [FILTER_TYPE_DUE_DATE]?: Date | null;
-  [FILTER_TYPE_CREATED_AT]?: Date | null;
-  [FILTER_TYPE_OVERDUE]?: boolean | null;
-};
+export type FilterTypes = Partial<{
+  [FILTER_TYPE_SEARCH]: string | null;
+  [FILTER_TYPE_STATUS]: components['schemas']['PayableStateEnum'] | null;
+  [FILTER_TYPE_DUE_DATE]: Date | null;
+  [FILTER_TYPE_CREATED_AT]: Date | null;
+  [FILTER_TYPE_SUMMARY_CARD]: PayablesTabFilter | null;
+}>;
 
 export type FilterValue =
   | components['schemas']['PayableStateEnum']
@@ -27,3 +28,24 @@ export type FilterValue =
   | Date
   | string
   | null;
+
+//TODO: better to map it with schema.json keyof values
+export type FieldValueTypes =
+  | 'document_id'
+  | 'counterpart_id'
+  | 'created_at'
+  | 'issued_at'
+  | 'due_date'
+  | 'status'
+  | 'amount'
+  | 'pay';
+
+export interface MonitePayableTableProps {
+  isShowingSummaryCards?: boolean;
+  fieldOrder?: Array<keyof FieldValueTypes>;
+  summaryCardFilters?: Record<string, PayablesTabFilter | null>;
+}
+
+export type PayablesTabFilter = NonNullable<
+  API['payables']['getPayables']['types']['parameters']['query']
+>;