+++ packages/sdk-react/src/components/receivables/InvoiceDetails/ExistingInvoiceDetails/components/EmailInvoiceDetails.tsx
@@ -3,12 +3,12 @@ import {
   useCallback,
   useEffect,
   useId,
+  useMemo,
   useState,
 } from 'react';
 import {
   Control,
   Controller,
-  SetFieldValue,
   useForm,
   UseFormGetValues,
 } from 'react-hook-form';
@@ -18,14 +18,17 @@ import { DefaultEmail } from '@/components/counterparts/CounterpartDetails/Count
 import { useMoniteContext } from '@/core/context/MoniteContext';
 import { MoniteScopedProviders } from '@/core/context/MoniteScopedProviders';
 import { useRootElements } from '@/core/context/RootElementsProvider';
-import { useFormPersist } from '@/core/hooks/useFormPersist';
-import { useReceivableContacts } from '@/core/queries';
+import {
+  useMe,
+  useMyEntity,
+  useReceivableById,
+  useReceivableContacts,
+} from '@/core/queries';
 import {
   useIssueReceivableById,
   useReceivableEmailPreview,
   useSendReceivableById,
 } from '@/core/queries/useReceivables';
-import i18n from '@/mocks/i18n';
 import { CenteredContentBox } from '@/ui/box';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { t } from '@lingui/macro';
@@ -59,30 +62,95 @@ interface EmailInvoiceDetailsProps {
   invoiceId: string;
   onClose: () => void;
 }
+interface EmailInvoiceFormProps extends EmailInvoiceDetailsProps {
+  subject: string;
+  body: string;
+  to: string;
+  isLoading: boolean;
+}
 
-export const EmailInvoiceDetails = (props: EmailInvoiceDetailsProps) => (
-  <MoniteScopedProviders>
-    <EmailInvoiceDetailsBase {...props} />
-  </MoniteScopedProviders>
-);
+export const EmailInvoiceDetails = (props: EmailInvoiceDetailsProps) => {
+  const { i18n } = useLingui();
+  const { data: me, isLoading: isLoadingUser } = useMe();
+  const { data: receivable, isLoading: isLoadingReceivable } =
+    useReceivableById(props.invoiceId);
+  const { data: contacts, isLoading: isLoadingContacts } =
+    useReceivableContacts(props.invoiceId);
+  const { entityName, isLoading: isLoadingEntity } = useMyEntity();
+
+  const defaultContact =
+    contacts && contacts.length
+      ? contacts?.find((c) => c.is_default) || contacts?.[0]
+      : undefined;
+
+  const to = defaultContact?.email ?? '';
+
+  const body =
+    defaultContact && me
+      ? t(i18n)`Hi ${defaultContact.first_name},
+
+          Please find the invoice attached as discussed.
+
+          Kind Regards,
+          ${me.first_name}`
+      : '';
+
+  const subject =
+    receivable && entityName
+      ? receivable.document_id
+        ? t(i18n)`Invoice ${receivable.document_id} from ${entityName}`
+        : t(i18n)`Invoice from ${entityName}`
+      : '';
+  const isLoading =
+    isLoadingContacts ||
+    isLoadingReceivable ||
+    isLoadingUser ||
+    isLoadingEntity;
+
+  return (
+    <MoniteScopedProviders>
+      <EmailInvoiceDetailsBase
+        {...props}
+        to={to}
+        body={body}
+        subject={subject}
+        isLoading={isLoading}
+      />
+    </MoniteScopedProviders>
+  );
+};
 
 const EmailInvoiceDetailsBase = ({
   invoiceId,
+  subject,
+  body,
+  to,
+  isLoading,
   onClose,
-}: EmailInvoiceDetailsProps) => {
+}: EmailInvoiceFormProps) => {
   const { i18n } = useLingui();
   const { monite, api } = useMoniteContext();
-  const { control, handleSubmit, getValues, setValue, trigger } = useForm({
+
+  const { control, handleSubmit, getValues, trigger, reset } = useForm({
     resolver: yupResolver(getEmailInvoiceDetailsSchema(i18n)),
-    defaultValues: {
-      subject: '',
-      body: '',
-      to: '',
-      // TODO: add support for multiple recipients, cc and bcc fields
-    },
+    defaultValues: useMemo(
+      () => ({
+        subject,
+        body,
+        to,
+        // TODO: add support for multiple recipients, cc and bcc fields
+      }),
+      [subject, body, to]
+    ),
   });
+
+  useEffect(() => {
+    reset({ subject, body, to });
+  }, [body, reset, subject, to]);
+
   // Use the same storage key for all invoices to avoid overloading the localStorage with dozens of saved form states
-  useFormPersist(`Monite-EmailInvoiceDetails-FormState`, getValues, setValue);
+  // TODO: Form persistance disabled as requested by Joao on Sep 5
+  // useFormPersist(`Monite-EmailInvoiceDetails-FormState`, getValues, setValue);
   const sendMutation = useSendReceivableById(invoiceId);
   const issueMutation = useIssueReceivableById(invoiceId);
 
@@ -243,7 +311,7 @@ const EmailInvoiceDetailsBase = ({
                     color="primary"
                     type="button"
                     form={formName}
-                    disabled={isDisabled}
+                    disabled={isDisabled || isLoading}
                     onClick={async () => {
                       const isValid = await trigger();
                       if (isValid) setPresentation(FormPresentation.Preview);
@@ -255,7 +323,7 @@ const EmailInvoiceDetailsBase = ({
                   color="primary"
                   type="submit"
                   form={formName}
-                  disabled={isDisabled}
+                  disabled={isDisabled || isLoading}
                 >{t(i18n)`Issue and send`}</Button>
               </Stack>
             </Grid>
@@ -269,16 +337,20 @@ const EmailInvoiceDetailsBase = ({
           p: isPreview ? 0 : '0 32px 32px 32px',
         }}
       >
-        {presentation == FormPresentation.Edit && (
-          <Form
-            invoiceId={invoiceId}
-            formName={formName}
-            handleIssueAndSend={handleIssueAndSend}
-            control={control}
-            isDisabled={isDisabled}
-            getValues={getValues}
-            setValue={setValue}
-          />
+        {isLoading ? (
+          <CenteredContentBox className="Monite-LoadingPage">
+            <CircularProgress />
+          </CenteredContentBox>
+        ) : (
+          presentation == FormPresentation.Edit && (
+            <Form
+              invoiceId={invoiceId}
+              formName={formName}
+              handleIssueAndSend={handleIssueAndSend}
+              control={control}
+              isDisabled={isDisabled}
+            />
+          )
         )}
         {isPreview && <Preview invoiceId={invoiceId} getValues={getValues} />}
       </DialogContent>
@@ -293,31 +365,13 @@ interface FormProps {
 }
 
 const RecipientSelector = ({
-  field,
   invoiceId,
   control,
-  getValues,
-  setValue,
 }: {
-  field: keyof FormProps;
   invoiceId: string;
   control: Control<FormProps>;
-  getValues: () => { [_: string]: any };
-  setValue: SetFieldValue<any>;
 }) => {
   const { data: contacts, isLoading } = useReceivableContacts(invoiceId);
-
-  useEffect(() => {
-    const currentValue = getValues()[field];
-    if (!currentValue && contacts && contacts.length > 0) {
-      const defaultContact = contacts.find((c) => c.is_default) ?? contacts[0];
-      setValue(field, defaultContact.email, {
-        shouldValidate: false,
-        shouldDirty: true,
-        shouldTouch: true,
-      });
-    }
-  }, [contacts, field, getValues, setValue]);
 
   const { root } = useRootElements();
 
@@ -325,7 +379,7 @@ const RecipientSelector = ({
 
   return (
     <Controller
-      name={field}
+      name="to"
       control={control}
       render={({ field, fieldState: { error } }) => (
         <FormControl
@@ -361,17 +415,15 @@ const Form = ({
   handleIssueAndSend,
   control,
   isDisabled,
-  getValues,
-  setValue,
 }: {
   invoiceId: string;
   formName: string;
   handleIssueAndSend: (e: BaseSyntheticEvent) => void;
   control: Control<FormProps>;
-  getValues: () => { [_: string]: any };
-  setValue: SetFieldValue<any>;
   isDisabled: boolean;
 }) => {
+  const { i18n } = useLingui();
+
   return (
     <form id={formName} noValidate onSubmit={handleIssueAndSend}>
       <Card sx={{ mb: 3 }}>
@@ -387,13 +439,7 @@ const Form = ({
             <Typography variant="body2" sx={{ minWidth: '52px' }}>{t(
               i18n
             )`To`}</Typography>
-            <RecipientSelector
-              field="to"
-              invoiceId={invoiceId}
-              control={control}
-              getValues={getValues}
-              setValue={setValue}
-            />
+            <RecipientSelector invoiceId={invoiceId} control={control} />
           </Stack>
         </CardContent>
       </Card>
@@ -461,6 +507,7 @@ const Preview = ({
   invoiceId: string;
   getValues: UseFormGetValues<FormProps>;
 }) => {
+  const { i18n } = useLingui();
   const { subject, body } = getValues();
   const { isLoading, preview, error, refresh } = useReceivableEmailPreview(
     invoiceId,