+++ packages/sdk-react/src/components/receivables/QuotesTable/QuotesTable.tsx
@@ -15,6 +15,7 @@ import { ReceivableCursorFields } from '@/enums/ReceivableCursorFields';
 import { CounterpartCellById } from '@/ui/CounterpartCell';
 import { DataGridEmptyState } from '@/ui/DataGridEmptyState';
 import { GetNoRowsOverlay } from '@/ui/DataGridEmptyState/GetNoRowsOverlay';
+import { DueDateCell } from '@/ui/DueDateCell';
 import {
   TablePagination,
   useTablePaginationThemeDefaultPageSize,
@@ -161,6 +162,7 @@ const QuotesTableBase = ({
         headerName: t(i18n)`Due date`,
         width: 120,
         valueFormatter: (value) => value && i18n.date(value, dateFormat),
+        renderCell: (params) => <DueDateCell data={params.row} />,
       },
       {
         field: 'status',