+++ packages/sdk-themes/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @team-monite/sdk-themes
 
+## 0.2.0-beta.3
+
+### Minor Changes
+
+- 59ff0c92: feat: Summary cards added for payables with theme customisation
+- 202101f0: feat(DEV-12699): disable summary cards via theme customisation
+- 0f7e21a5: feat: Move order of columns - extend config to pass order
+- 84ee5502: feat(DEV-12179): custom Tabs for Payables
+
+### Patch Changes
+
+- 43a2d4e6: encapsulate entity check logic
+
+## 0.2.0-beta.2
+
+### Minor Changes
+
+- 0a37dc68: Add country and city name next to counterpart name in tables.
+- 6c3faf43: Add coloured avatars for counterparts in tables, improve look of Counterparts table.
+- cddff2af: Fix table filters layout, fix 'Status' filter font size
+- 5d4c1274: Add the capability to change invoice recipient and preview invoice email before sending.
+
 ## 0.2.0-beta.1
 
 ### Minor Changes